mon: disable gmt_hitset if not supported
the gmt_hitset is enabled by default in the ctor of pg_pool_t, this
is intentional. because we want to remove this setting and make
gmt_hitset=true as a default in future. but this forces us to
disable it explicitly when preparing a new pool if any OSD does
not support gmt hitset.

Fixes: #12968
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit e48cec3dc93b3988dcd8924933deb1b3a43e1d0f)

Conflicts:
	src/mon/OSDMonitor.cc
		remove unrelated hunk